style(widget-schema-helper):  separate less relevant code into functions
